% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_note_wb.R
\name{add_note_wb}
\alias{add_note_wb}
\title{Add a note or source}
\usage{
add_note_wb(noteTitle = "", note = "")
}
\arguments{
\item{noteTitle}{The title of the note, usually "Note:" or "Source:"}

\item{note}{The content of the note}
}
\value{
None
}
\description{
Add a note or source to the bottom of a plot
}
\examples{
add_note_wb(noteTitle = "Source:", note = "World Bank")
}
