% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes_barchart_wb_DEPRECATED.R
\name{axes_barchart_wb}
\alias{axes_barchart_wb}
\title{Style axes for bar charts, based on the World Bank data visualization style guide}
\usage{
axes_barchart_wb(xExpansion = 0)
}
\arguments{
\item{xExpansion}{Number to exand the right side of the x scale, to give room for data labels}
}
\value{
None
}
\description{
Add this function to a ggplot2 object to style the axes of a horizontal bar chart
}
\examples{
axes_barchart_wb()
}
