% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes_wb_DEPRECATED.R
\name{axes_wb}
\alias{axes_wb}
\title{Style axes based on the World Bank data visualization style guide}
\usage{
axes_wb(
  extendYToZero = FALSE,
  addExpansion = FALSE,
  addZeroLine = FALSE,
  hideVerticalGrid = FALSE,
  hideHorizontalGrid = FALSE
)
}
\arguments{
\item{extendYToZero}{Boolean to extend the y limits to include zero}

\item{addExpansion}{Boolean to expand the scales}

\item{addZeroLine}{Boolean to add a zero line for the Y scale. This will also extend the Y scale to zero}

\item{hideVerticalGrid}{Boolean to hide the vertical grid lines (mostly for datetime x axes)}

\item{hideHorizontalGrid}{Boolean to hide the horizontal grid lines (mostly for bar charts and beeswarm plots)}

\item{barChart}{Boolean to style axes for bar charts (moves to x axis to the top, and capitalizes y axis labels)}
}
\value{
None
}
\description{
Add this function to a ggplot2 object to
\itemize{
\item extend the y limits to include zero
\item remove the expansion on the x and y scales
\item add a horizontal line at y = 0
}
}
\examples{
axes_wb(extendYToZero = TRUE, addExpansion = TRUE, addZeroLine = TRUE)
}
