% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country data}
\format{
Data frame of available countries and regions from the World Bank API, as returned by wbstats::wbcountries
}
\source{
https://api.worldbank.org/v2/country
}
\usage{
countries
}
\description{
Country data
}
\keyword{datasets}
