% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color.R
\name{get_color}
\alias{get_color}
\title{Get the color from its name}
\usage{
get_color(colorName)
}
\arguments{
\item{colorName}{The name of the color}
}
\value{
Hex code of the color
}
\description{
Get the color from its name
}
\examples{
get_color('darkText')

}
