% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_font_weight.R
\name{get_font_weight}
\alias{get_font_weight}
\title{Get the font weight from the font face}
\usage{
get_font_weight(fontFace)
}
\arguments{
\item{fontFace}{The font face name}
}
\value{
The font weight
}
\description{
Get the font weight from the font face
}
\examples{
get_font_weight('semibold')

}
