% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_wb.R
\name{ggsave_wb}
\alias{ggsave_wb}
\title{Save plots with good default dimensions}
\usage{
ggsave_wb(
  filename = "my-worldbank-plot.png",
  units = "px",
  width = 960,
  height = 540,
  scale = 1,
  dpi = 200,
  ...
)
}
\arguments{
\item{filename}{The name of your image file to save}

\item{units}{Units of the dimensions of the imagefile}

\item{width}{The width of the image file}

\item{height}{The height of the image file}

\item{scale}{The scale factor}

\item{...}{Other arguments passed to ggsave()}
}
\value{
None
}
\description{
Save plots with good default dimensions
}
\examples{
\dontrun{
ggsave_wb(filename = "name-of-your-plot.jpg")
}

}
