% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{life.expectancy}
\alias{life.expectancy}
\title{Life expactancy data}
\format{
time series data for the SP.DYN.LE00.IN indicator for all countries and regions, as returned by wbstats::wb_data("SP.DYN.LE00.IN")
}
\source{
https://data.worldbank.org/indicator/SP.DYN.LE00.IN
}
\usage{
life.expectancy
}
\description{
Life expactancy data
}
\keyword{datasets}
