% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_wb_d.R
\name{scale_color_wb_d}
\alias{scale_color_wb_d}
\title{Discrete colour scales with WB palettes}
\usage{
scale_color_wb_d(..., palette = "default", na.value = WBCOLORS[["noData"]])
}
\arguments{
\item{...}{Other arguments passed to scale_colour_discrete()}

\item{palette}{The discrete WB color palette to use. One of 'default', 'defaultText', 'region', 'regionText', 'income', 'gender', 'urbanisation', 'age', or 'binary'}

\item{na.value}{Color to use for NA values}
}
\value{
None
}
\description{
Discrete colour scales with WB palettes
}
\examples{
scale_color_wb_d(palette = "default")

}
