% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_binned_wb.R
\name{scale_fill_binned_wb}
\alias{scale_fill_binned_wb}
\title{Binned fill scales with WB palettes}
\usage{
scale_fill_binned_wb(..., palette = "seq", na.value = WBCOLORS[["noData"]])
}
\arguments{
\item{...}{Other arguments passed to scale_fill_binned()}

\item{palette}{The WB color palette to use. One of 'seq', 'seqRev', 'seqB', 'seqY', 'seqP', 'divPosNeg', or 'divLR'}

\item{na.value}{Color to use for NA values}
}
\value{
None
}
\description{
Binned fill scales with WB palettes
}
\examples{
scale_fill_binned_wb(palette = "seq", n.breaks = 8)

}
