% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_wb_b.R, R/scale_fill_wb_d.R
\name{scale_fill_wb_d}
\alias{scale_fill_wb_d}
\title{Binned fill scales with WB palettes}
\usage{
scale_fill_wb_d(..., palette = "default", na.value = WBCOLORS[["noData"]])

scale_fill_wb_d(..., palette = "default", na.value = WBCOLORS[["noData"]])
}
\arguments{
\item{...}{Other arguments passed to scale_colour_discrete()}

\item{palette}{The discrete WB color palette to use. One of 'default', 'defaultText', 'region', 'regionText', 'income', 'gender', 'urbanisation', 'age', or 'binary'}

\item{na.value}{Color to use for NA values}
}
\value{
None

None
}
\description{
Binned fill scales with WB palettes

Discrete fill scales with WB palettes
}
\examples{
scale_fill_wb_d(palette = "seq", n.breaks = 8)

scale_fill_wb_d(palette = "default")

}
