% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_shape_wb.R
\name{scale_shape_wb}
\alias{scale_shape_wb}
\title{Discrete shape scale, with filled shapes}
\usage{
scale_shape_wb(...)
}
\arguments{
\item{...}{Arguments passed to scale_shape_manual()}
}
\value{
None
}
\description{
Discrete shape scale, with filled shapes
}
\examples{
scale_shape_wb()

}
