% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wb_dark_blue}
\alias{wb_dark_blue}
\title{World Bank dark blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
wb_dark_blue
}
\description{
World Bank dark blue
}
\seealso{
Other World Bank corporate color palette: 
\code{\link{wb_accent}},
\code{\link{wb_dark_green}},
\code{\link{wb_grey}},
\code{\link{wb_light_blue}},
\code{\link{wb_light_green}}
}
\concept{World Bank corporate color palette}
\keyword{datasets}
