% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wb_grey}
\alias{wb_grey}
\title{World Bank grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
wb_grey
}
\description{
World Bank grey
}
\seealso{
Other World Bank corporate color palette: 
\code{\link{wb_accent}},
\code{\link{wb_dark_blue}},
\code{\link{wb_dark_green}},
\code{\link{wb_light_blue}},
\code{\link{wb_light_green}}
}
\concept{World Bank corporate color palette}
\keyword{datasets}
