% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wb_palette}
\alias{wb_palette}
\title{World Bank color palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
wb_palette
}
\description{
Color palette dervied from World Bank's corporate colors.
}
\keyword{datasets}
