% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{wb_theme}
\alias{wb_theme}
\title{Set World Bank theme}
\usage{
wb_theme(
  font_family_base = choose_fonts()$base,
  font_family_title = choose_fonts()$title,
  font_size = 12,
  ...
)
}
\arguments{
\item{font_family_base}{Character. Base font family name.
Defaults to choice stored in \code{choose_fonts()$base}}

\item{font_family_title}{Character. Font family for the title.
Defaults to choice stored in \code{choose_fonts()$title}}

\item{font_size}{Numeric. Base font size. Defaults to 12.}

\item{...}{Passes theme parameters to \code{ggplot2::theme()}}
}
\description{
Core wbplot theme that is modified by other themes.
}
