% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{wb_theme_bar}
\alias{wb_theme_bar}
\title{Provide a bar plot theme}
\usage{
wb_theme_bar(
  font_family_base = choose_fonts()$base,
  font_family_title = choose_fonts()$title,
  font_size = 12,
  x_grid = TRUE,
  x_axis = FALSE,
  use_axis_title = TRUE,
  ...
)
}
\arguments{
\item{font_family_base}{Character. Base font family name.
Defaults to choice stored in \code{choose_fonts()$base}}

\item{font_family_title}{Character. Font family for the title.
Defaults to choice stored in \code{choose_fonts()$title}}

\item{font_size}{Numeric. Base font size. Defaults to 12.}

\item{x_grid}{Boolean. Whether to include x axis grid lines and
any x axis text.}

\item{x_axis}{Boolean. Whether to show x axis scale, title, and labels.
If so, include x axis title and label text.
If not, remove x axis title and label text.}

\item{use_axis_title}{Boolean. Whether to have axis title.
If so, color title black and label text grey.
If not, color label text grey.}

\item{...}{Passes theme parameters to \code{ggplot2::theme()}}
}
\description{
Removes y axis grid lines
}
