% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{WBPALETTES}
\alias{WBPALETTES}
\title{World Bank color palettes}
\format{
A named list with the sequential, diverging and categorical color palettes, derived from data-raw/wb-colors.json file
}
\usage{
WBPALETTES
}
\description{
World Bank color palettes
}
\keyword{datasets}
