% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{WBSTYLE}
\alias{WBSTYLE}
\title{World Bank data visualization style}
\format{
A nested list with the World Bank visualization styling, derived from ingesting data-raw/wb-style.json file
}
\usage{
WBSTYLE
}
\description{
World Bank data visualization style
}
\keyword{datasets}
