% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axes_wb.R
\name{axes_wb}
\alias{axes_wb}
\title{Style axes based on the World Bank data visualization style guide}
\usage{
axes_wb(extendYToZero = FALSE, addExpansion = FALSE, addZeroLine = TRUE)
}
\arguments{
\item{extendYToZero}{Boolean to extend the y limits to include zero}

\item{addExpansion}{Boolean to expand the scales}

\item{addZeroLine}{Boolean to add a zero line for the Y scale. This will also extend the Y scale to zero}
}
\value{
None
}
\description{
Add this function to a ggplot2 object to
\itemize{
\item extend the y limits to include zero
\item remove the expansion on the x and y scales
\item add a horizontal line at y = 0
}
}
\examples{
axes_wb(extendYToZero = TRUE, addExpansion = TRUE, addZeroLine = TRUE)
}
