% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_wb_b.R
\name{scale_color_wb_b}
\alias{scale_color_wb_b}
\title{Binned color scales with WB palettes}
\usage{
scale_color_wb_b(
  ...,
  palette = "seq",
  direction = 1,
  na.value = WBCOLORS[["noData"]]
)
}
\arguments{
\item{...}{Other arguments passed to scale_fill_binned()}

\item{palette}{The WB color palette to use. One of 'seq', 'seqRev', 'seqB', 'seqY', 'seqP', 'divPosNeg', or 'divLR'}

\item{direction}{Sets the order of colours in the scale. 1 is the default,  if -1, the order of colours is reversed}

\item{na.value}{Color to use for NA values}
}
\value{
None
}
\description{
Binned color scales with WB palettes
}
\examples{
scale_color_wb_b(palette = "seq", n.breaks = 8)

}
