% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\name{theme_wb}
\alias{theme_wb}
\title{ggplot theme for World Bank visualizations}
\usage{
theme_wb(
  chartType = "",
  xExpansion = 0,
  addXZeroLine = FALSE,
  addYZeroLine = FALSE
)
}
\arguments{
\item{chartType}{Activates chart specific styling. Should be one of "line", "bar", "beeswarm" or "scatter"}

\item{xExpansion}{Expand the x axis to make room for data labels on bar charts}

\item{addXZeroLine}{Boolean for adding a line for zero on the X axis}

\item{addYZeroLine}{Boolean for adding a line for zero on the Y axis}
}
\value{
None
}
\description{
Creates overrides for a ggplot theme, using the World Bank data visualization style
}
\examples{
theme_wb()

}
