% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\name{theme_wb}
\alias{theme_wb}
\title{ggplot theme for World Bank visualizations}
\usage{
theme_wb()
}
\value{
None
}
\description{
Creates overrides for a ggplot theme, using the World Bank data visualization style
}
\examples{
theme_wb()

}
